requirements.txt
loguru==0.7.2
python-dotenv==1.0.1
requests==2.31.0
